+++ Changelog.md
@@ -55,6 +55,7 @@
 * Redone parsing `CommenAuthors`
 * Remove usage of deprecated `Coordinates#get_column_number`
 * Use instance method `Relationships#parse_file` instead of class method `Relationships.parse_rels`
+* Simplify parsing `DocxParagraphRun#font` for `asciiTheme`
   
 ### Fixes
 * Fix crash on empty coordinates list of chart